Spotify_Music_Analysis_Project/requirements.txt
# Professional Spotify Music Analysis - Requirements

# Core Data Science Libraries
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Spotify API and Web Requests
spotipy>=2.23.0
requests>=2.31.0

# Database and SQL
sqlite3  # Built-in with Python
sqlalchemy>=2.0.0

# Data Processing and Analysis
scipy>=1.10.0
scikit-learn>=1.3.0
statsmodels>=0.14.0

# Configuration and Environment
python-dotenv>=1.0.0
pyyaml>=6.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
tqdm>=4.65.0
colorama>=0.4.6
tabulate>=0.9.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
